Finished the cleaning script, and added wrappers
The wrappers are intended to be run from cron. They output errors to
stdout so it will be captured by cron to be sent by email. All output
is also redirected to a log file.

The cleaning script is intended to called with a parameter ('week',
'month' 'year') and run at that interval. So if you pass "week" to the
script, it assumes that it was a week ago it was run.

If called with eg 'week', then it will take today's backup and cp-lax
it to the source's weekly folder. Then it will delete all but the last
5 backups in the weekly folder (so there is one snapshot from every
week, going back five weeks).

A similar pattern is applied for day, month, week and year. Details
here:

Interval   cp-lax   delete   num-to-save
day        no       yes      14
week       yes      yes      5
month      yes      yes      12
year       yes      no       -